Fix bad merge on integration-tests

+++ java/api-test/integration-tests/src/test/java/com/microsoft/semantickernel/connectors/memory/postgresql/PostgreSQLMemoryStoreTest.java
@@ -409,7 +409,7 @@ public class PostgreSQLMemoryStoreTest {
         String collection = setUpNearestMatches();
 
         // Act
-        double threshold = -1;
+        float threshold = -1f;
         Collection<Tuple2<MemoryRecord, Float>> topNResults =
                 db.getNearestMatchesAsync(collection, compareEmbedding, topN, threshold, false)
                         .block();
@@ -432,7 +432,7 @@ public class PostgreSQLMemoryStoreTest {
         String collection = setUpNearestMatches();
 
         // Act
-        double threshold = -1;
+        float threshold = -1f;
         Collection<Tuple2<MemoryRecord, Float>> topNResults =
                 db.getNearestMatchesAsync(collection, compareEmbedding, 0, threshold, false)
                         .block();
@@ -449,7 +449,7 @@ public class PostgreSQLMemoryStoreTest {
         String collection = setUpNearestMatches();
 
         // Act
-        double threshold = -1;
+        float threshold = -1f;
         Collection<Tuple2<MemoryRecord, Float>> topNResults =
                 db.getNearestMatchesAsync(collection, compareEmbedding, 0, threshold, false)
                         .block();
@@ -468,7 +468,7 @@ public class PostgreSQLMemoryStoreTest {
         db.createCollectionAsync(collection).block();
 
         // Act
-        double threshold = -1;
+        float threshold = -1f;
         Collection<Tuple2<MemoryRecord, Float>> topNResults =
                 db.getNearestMatchesAsync(
                                 collection, compareEmbedding, Integer.MAX_VALUE, threshold, false)
@@ -488,7 +488,7 @@ public class PostgreSQLMemoryStoreTest {
         String collection = setUpNearestMatches();
 
         // Act
-        double threshold = 0.75;
+        float threshold = 0.75f;
         Tuple2<MemoryRecord, Float> topNResultDefault =
                 db.getNearestMatchAsync(collection, compareEmbedding, threshold, false).block();
         Tuple2<MemoryRecord, Float> topNResultWithEmbedding =
@@ -513,7 +513,7 @@ public class PostgreSQLMemoryStoreTest {
         String collection = setUpNearestMatches();
 
         // Act
-        double threshold = 0.75;
+        float threshold = 0.75f;
         Tuple2<MemoryRecord, Float> topNResult =
                 db.getNearestMatchAsync(collection, compareEmbedding, threshold, false).block();
 
@@ -533,7 +533,7 @@ public class PostgreSQLMemoryStoreTest {
         db.createCollectionAsync(collection).block();
 
         // Act
-        double threshold = -1;
+        float threshold = -1f;
         Tuple2<MemoryRecord, Float> topNResults =
                 db.getNearestMatchAsync(collection, compareEmbedding, threshold, false).block();
 
@@ -565,7 +565,7 @@ public class PostgreSQLMemoryStoreTest {
 
         // Act
         Collection<Tuple2<MemoryRecord, Float>> topNResults =
-                db.getNearestMatchesAsync(collection, compareEmbedding, topN, 0.75, true).block();
+                db.getNearestMatchesAsync(collection, compareEmbedding, topN, 0.75f, true).block();
         Collection<String> topNKeys =
                 topNResults.stream()
                         .map(tuple -> tuple.getT1().getKey())
